refactor(sdk): flatten test directory structure
- Consolidate actions into `test/actions.ts`
- Flatten `account`, `group`, and `treasury` tests into root `test/` folder
- Update imports in tests and helpers
- Remove subdirectories
